program working for decimal numbers for simple expression

+++ src/cli/exprEvalMain.java
@@ -6,7 +6,7 @@ import ExprEval.evalExpr;
 public class exprEvalMain {
     public static void main(String[] args) {
         evalExpr evaluateExpression = new evalExpr();
-        int result = evaluateExpression.evaluate(args[0]);
+        double result = evaluateExpression.evaluate(args[0]);
         System.out.println(result);
     }
 }
